// apps/api/src/images/images.types.ts
export interface ImageModel {
  publicId: string;
  order: number;
  thumbnailUrl: string;
  thumbnailWidth: number;
  thumbnailHeight: number;
  thumbnailSize: number;
  mediumUrl: string;
  mediumWidth: number;
  mediumHeight: number;
  mediumSize: number;
  originalUrl: string;
  originalWidth: number;
  originalHeight: number;
  originalSize: number;
}